Fix - 상태관리 App 컴포넌트로 옮기기
