+++ src/app.jsx
@@ -1,7 +1,22 @@
+import { useState } from 'react';
 import Page from './page';
 
 export default function App() {
+  const [state, setState] = useState({
+    count: 0,
+  });
+
+  const { count } = state;
+
+  function handleClick(value) {
+    setState({
+      count: count + value,
+    });
+  }
   return (
-    <Page />
+    <Page
+      count={count}
+      onClick={handleClick}
+    />
   );
 }
