+++ src/page.jsx
@@ -1,29 +1,17 @@
-import { useState } from 'react';
 import Counter from './counter';
 import Buttons from './buttons';
 
-export default function Page() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-
-  const { count } = state;
-
-  function handleClick(value) {
-    setState({
-      count: count + value,
-    });
-  }
+export default function Page({ count, onClick }) {
   return (
     <div>
       <p>Counter</p>
       <Counter
         count={count}
-        onClick={handleClick}
+        onClick={onClick}
       />
       <Buttons
         count={count}
-        onClick={handleClick}
+        onClick={onClick}
       />
     </div>
   );
